add api succesfuly responses

+++ src/main/java/org/example/incidentmanagement/service/impl/UserRoleServiceImpl.java
@@ -1,11 +1,10 @@
 package org.example.incidentmanagement.service.impl;
 
 import org.example.incidentmanagement.dto.requests.CreateUserRoleRequestDto;
-import org.example.incidentmanagement.dto.requests.UpdateUserRoleRequestDto;
 import org.example.incidentmanagement.dto.response.UserRoleResponseDto;
 import org.example.incidentmanagement.entity.UserRoles;
 import org.example.incidentmanagement.exceptions.CustomException;
-import org.example.incidentmanagement.exceptions.ErrorCodes;
+import org.example.incidentmanagement.exceptions.ResponseCodes;
 import org.example.incidentmanagement.mappers.UserRoleMapper;
 import org.example.incidentmanagement.repository.UserRolesRepository;
 import org.example.incidentmanagement.service.RoleService;
@@ -69,7 +68,7 @@ public class UserRoleServiceImpl implements UserRoleService {
     public UserRoleResponseDto createUserRole(CreateUserRoleRequestDto createUserRoleRequestDto) {
         if (createUserRoleRequestDto == null){
             logger.info("Create user role with null request");
-            throw new CustomException(ErrorCodes.NULL_CODE);
+            throw new CustomException(ResponseCodes.NULL_CODE);
         }
 
         logger.info("Called Create user role with request: {}, {}, {}", createUserRoleRequestDto.getUserId(),
@@ -81,18 +80,18 @@ public class UserRoleServiceImpl implements UserRoleService {
 
         if (!roleService.existsRole(createUserRoleRequestDto.getRoleId())) {
             logger.info("Role with id {} not exists", createUserRoleRequestDto.getRoleId());
-            throw new CustomException(ErrorCodes.INVALID_ROLE);
+            throw new CustomException(ResponseCodes.INVALID_ROLE);
         }
 
         if (!userService.existsUser(createUserRoleRequestDto.getUserId())) {
             logger.info("User with id {} not exists", createUserRoleRequestDto.getUserId());
-            throw new CustomException(ErrorCodes.INVALID_USER);
+            throw new CustomException(ResponseCodes.INVALID_USER);
         }
 
         if (existsUserRole(createUserRoleRequestDto.getUserId(), createUserRoleRequestDto.getRoleId())) {
             logger.info("User id '{}' already have role with id '{}' ", createUserRoleRequestDto.getUserId(),
                     createUserRoleRequestDto.getRoleId());
-            throw new CustomException(ErrorCodes.INVALID_USER);
+            throw new CustomException(ResponseCodes.INVALID_USER);
         }
 
         userRolesRepository.save(userRoles);
@@ -103,6 +102,13 @@ public class UserRoleServiceImpl implements UserRoleService {
     @Override
     public void deleteUserRole(int id) {
         logger.info("Deleting user role with id: " + id);
+        UserRoles userRoles = userRolesRepository.findById(id);
+
+        if (userRoles == null) {
+            logger.info("User role with id {} not exists", id);
+            throw new CustomException(ResponseCodes.INVALID_USER);
+        }
+
         userRolesRepository.deleteById(id);
     }
